Simplify marketing pages without force dynamic
